test: enable DocumentCommandHandlers-integ-test
