+++ test/spec/CSSUtils-test.js
@@ -19,7 +19,7 @@
  *
  */
 
-/*global describe, it, expect, beforeEach, afterEach, beforeAll, afterAll, awaitsForDone, jasmine, jsPromise */
+/*global describe, it, expect, beforeEach, afterEach, awaitsForDone */
 
 define(function (require, exports, module) {
 
@@ -661,7 +661,6 @@ define(function (require, exports, module) {
             });
         });
     }); // describe("CSSUtils")
-
 
 
     describe("CSS Parsing", function () {
@@ -1800,105 +1799,6 @@ define(function (require, exports, module) {
 
     }); //describe("CSS Parsing")
 
-    describe("integration:CSS Integration Tests", function () {
-        // todo not working
-
-        let testPath = SpecRunnerUtils.getTestPath("/spec/CSSUtils-test-files"),
-            testWindow,
-            CSSUtils,
-            DocumentManager,
-            FileViewController;
-
-        beforeAll(async function () {
-            testWindow = await SpecRunnerUtils.createTestWindowAndRun();
-            // Load module instances from brackets.test
-            CSSUtils            = testWindow.brackets.test.CSSUtils;
-            DocumentManager     = testWindow.brackets.test.DocumentManager;
-            FileViewController  = testWindow.brackets.test.FileViewController;
-            // Load test project
-            await SpecRunnerUtils.loadProjectInTestWindow(testPath);
-        }, 20000);
-
-        afterAll(async function () {
-            CSSUtils            = null;
-            DocumentManager     = null;
-            FileViewController  = null;
-            await SpecRunnerUtils.closeTestWindow();
-        });
-
-        afterEach(function () {
-            testWindow.closeAllFiles();
-        });
-
-
-        describe("Working with real public CSSUtils API", function () {
-
-            it("should include comment preceding selector (issue #403)", async function () {
-                let rules = await jsPromise(CSSUtils.findMatchingRules("#issue403"));
-                expect(rules.length).toBe(1);
-                expect(rules[0].lineStart).toBe(4);
-                expect(rules[0].lineEnd).toBe(7);
-            }, 20000);
-
-            it("should continue search despite unreadable files (issue #10013)", async function () {
-                // Add a nonexistent CSS file to the ProjectManager.getAllFiles() result, which will force a file IO error
-                // when we try to read the file later. Similar errors may arise in real-world for non-UTF files, etc.
-                SpecRunnerUtils.injectIntoGetAllFiles(testWindow, testPath + "/doesNotExist.css");
-
-                var promise = CSSUtils.findMatchingRules("html");
-                promise.done(function (result) {
-                    expect(result.length).toBeGreaterThan(0);
-                });
-                await awaitsForDone(promise, "CSSUtils.findMatchingRules()");
-            });
-        });
-
-        describe("Working with unsaved changes", function () {
-
-            it("should return the correct offsets if the file has changed", async function () {
-                var promise = FileViewController.openAndSelectDocument(testPath + "/simple.css", FileViewController.PROJECT_MANAGER);
-                await awaitsForDone(promise, "FileViewController.openAndSelectDocument()");
-
-                var rules = null;
-
-                var doc = DocumentManager.getCurrentDocument();
-
-                // Add several blank lines at the beginning of the text
-                doc.setText("\n\n\n\n" + doc.getText());
-
-                // Look for ".FIRSTGRADE"
-                var promise = CSSUtils.findMatchingRules(".FIRSTGRADE")
-                    .done(function (result) { rules = result; });
-                await awaitsForDone(promise);
-
-                doc = null;
-                expect(rules.length).toBe(1);
-                expect(rules[0].lineStart).toBe(16);
-                expect(rules[0].lineEnd).toBe(18);
-            });
-
-            it("should return a newly created rule in an unsaved file", async function () {
-                var promise = FileViewController.openAndSelectDocument(testPath + "/simple.css", FileViewController.PROJECT_MANAGER);
-                await awaitsForDone(promise, "FileViewController.openAndSelectDocument()");
-                var rules = null;
-
-                var doc = DocumentManager.getCurrentDocument();
-
-                // Add a new selector to the file
-                doc.setText(doc.getText() + "\n\n.TESTSELECTOR {\n    font-size: 12px;\n}\n");
-
-                // Look for the selector we just created
-                var promise = CSSUtils.findMatchingRules(".TESTSELECTOR")
-                    .done(function (result) { rules = result; });
-                await awaitsForDone(promise, "CSSUtils.findMatchingRules()");
-
-                doc = null;
-                expect(rules.length).toBe(1);
-                expect(rules[0].lineStart).toBe(24);
-                expect(rules[0].lineEnd).toBe(26);
-            });
-        });
-    });
 
     describe("CSSUtils - Other", function () {
         function doAddRuleTest(options) {
@@ -2048,7 +1948,6 @@ define(function (require, exports, module) {
     });
 
 
-
     // Unit Tests: "HTMLUtils (css)"
     describe("HTMLUtils InlineEditorProviders", function () {
         var editor;
@@ -2085,7 +1984,6 @@ define(function (require, exports, module) {
             });
         });
     });
-
 
 
     // These tests are based on the implementation spec at https://github.com/adobe/brackets/wiki/CSS-Context-API-implementation-spec.
@@ -2570,7 +2468,6 @@ define(function (require, exports, module) {
     });
 
 
-
     // These are tests related to Shapes editor requirements for determining the start/end range of a css property
     describe("CSS Context Info Ranges", function () {
 
@@ -2659,7 +2556,6 @@ define(function (require, exports, module) {
             });
         });
     });
-
 
 
     describe("CSS Regions", function () {
